fix: change write page to be index.vue and [id].vue
